pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["karinavolodina <volodinakarina27@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.9.6"
#python = "^3.10"
django = "^4.2.2"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"
rest-pandas = "^1.1.0"
django-filter = "^23.2"
drf-spectacular = "^0.26.3"
match = "^0.3.2"
pydantic = "^2.2.0"
pytest-factoryboy = "^2.5.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
ipython = "^8.14.0"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[toll.black]
line-lenght = 120
skip-string-normalization = true
skip-magic-trailing-coma = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | migrations
)/
'''

#[tool.isort]
#multi_line_output = 3
#include_trailing_comma = true
#line_lenght = 120
#combine_as_imports = true
#combine_star = true
#skip_glob = [
#    '**/migrations/**',
#    '.git',
#    '.venv'
#]
[tool.isort]
py_version=310
line_length = 120
multi_line_output = 3
include_trailing_comma = true
combine_as_imports = true
combine_star = true
force_grid_wrap = 0
skip = ['.gitignore', '.dockerignore']
skip_glob = [
    '**/migrations/**',
    '.git',
    '.venv',
    '.idea'
]
